requirements.txt
transformers
torch
tqdm
llama-cpp-python
huggingface-hub
fastapi
uvicorn
